Firestore read test extracted to a separate screen
